core/comms/secure_channel: Handle client null nonces as zero
This is probably not the change we want, but the testing change we need
right now. If a client connects and has no security modes configured for
OPC-UA it should be ok to provide a null bytestring as the nonce. We
will treat it as zero.

The empty bytestring will still be rejected with this change.